hotfix: groupRoom에 게스트 삭제 시에도 member_number 무조건 삭제되는 현상 수정

+++ src/main/java/challenge/nDaysChallenge/service/member/MemberService.java
@@ -111,21 +111,35 @@ public class MemberService {
         }
 
         for (GroupRoom groupRoom : groupRooms) {
-            groupRoom.deleteHostConnection();
+            if (groupRoom.getMember().equals(member)) {
+                groupRoom.deleteHostConnection();
+            } else {
+                groupRoom.deleteGuestConnection();
+            }
         }
 
         for (SingleRoom singleRoom:singleRooms){
             singleRoom.deleteConnection();
         }
+        em.flush();
+        em.clear();
 
-        //레포지토리에서 직접 엔티티 삭제
-        dajimRepository.deleteAll(dajims); //다짐 + 이모션 삭제
+        //레포지토리에서 직접 삭제
+        if (!dajims.isEmpty()){
+            dajimRepository.deleteAll(dajims); //탈퇴 회원 다짐 삭제 -> 이모션도 삭제
+        }
 
-        roomMemberRepository.deleteAll(roomMembers);
+        if (!roomMembers.isEmpty()){
+            roomMemberRepository.deleteAll(roomMembers); //탈퇴 회원 룸멤버 테이블에서 삭제
+        }
 
-        stampRepository.deleteAll(stamps);
+        if (!stamps.isEmpty()){
+            stampRepository.deleteAll(stamps); //탈퇴 회원 스탬프 삭제
+        }
 
-        singleRoomRepository.deleteAll(singleRooms);
+        if (!singleRooms.isEmpty()){
+            singleRoomRepository.deleteAll(singleRooms);
+        }
 
         //멤버 삭제
         memberRepository.delete(member);
